requirements.txt
black==23.7.0
bleach==6.0.0
build==0.10.0
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.6
commonmark==0.9.1
docutils==0.20.1
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.0.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pep517==0.13.0
pkginfo==1.9.6
platformdirs==3.10.0
Pygments==2.15.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyYAML==6.0.1
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
six==1.16.0
tomli==2.0.1
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==2.0.4
webencodings==0.5.1
zipp==3.16.2
